# docs/openapi/pipeline-jobs.openapi.yaml
openapi: 3.0.3
info:
  title: Market-Assistant — Pipeline & JD 数据集 API（子集）
  description: |
    与 `docs/pipeline-job-output-spec.md` 一致的核心路径；完整行为以 `pipeline/views.py` 为准。
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:8000
    description: 本地 Django 默认端口
...
paths:
  /api/jobs/:
    get:
      tags: [jobs]
      summary: 任务列表（最近 200 条）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineJob"
    post:
      tags: [jobs]
      summary: 创建任务并异步执行
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePipelineJob"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "503":
          description: LOW_GI_PROJECT_ROOT 未配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/jobs/{id}/:
    get:
      tags: [jobs]
      summary: 任务详情
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "404":
          description: Not found
    patch:
      tags: [jobs]
      summary: 仅更新 report_config（外部市场表等；预设关注词/场景已废弃的请求字段将被忽略）
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchReportConfig"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "400":
          description: 校验失败（未知键或体积过大等）
        "404":
          description: Not found
...
components:
  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
    CreatePipelineJob:
      type: object
      required: [platform, keyword]
      properties:
        platform:
          type: string
          example: jd
        keyword:
          type: string
        max_skus:
          type: integer
        page_start:
          type: integer
        page_to:
          type: integer
        pipeline_run_dir:
          type: string
        cookie_file_path:
          type: string
        cookie_text:
          type: string
        pvid:
          type: string
        request_delay:
          type: string
        list_pages:
          type: string
        scenario_filter_enabled:
          type: boolean
        report_config:
          $ref: "#/components/schemas/ReportConfig"
    ReportConfig:
      type: object
      description: |
        允许键以后端 ``validate_report_config_body`` 为准（含各 ``llm_*``、``chapter8_text_mining_probe*``、``external_market_table_rows`` 等）。
        已废弃：``comment_focus_words``、``comment_scenario_groups``、``llm_scenario_group_summaries``（请求中若出现将被忽略）。
      properties:
        external_market_table_rows:
          type: array
          items:
            type: object
            properties:
              indicator:
                type: string
              value_and_scope:
                type: string
              source:
                type: string
              year:
                type: string
    PatchReportConfig:
      type: object
      required: [report_config]
      properties:
        report_config:
          $ref: "#/components/schemas/ReportConfig"
    PipelineJob:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        keyword:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, cancelled]
        cancellation_requested:
          type: boolean
        run_dir:
          type: string
        error_message:
          type: string
        analysis_artifacts:
          type: object
          additionalProperties:
            type: boolean
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        report_config:
          $ref: "#/components/schemas/ReportConfig"
...